refactor(scene-engine): promote imageSignedUrlOptions to shared
