+++ frontend/src/features/scene-engine/steps/character/character.queries.ts
@@ -18,14 +18,3 @@ export const charactersOptions = (projectId: string, storyId: string) =>
     enabled: !!projectId && !!storyId,
     staleTime: Infinity,
   });
-
-export const imageSignedUrlOptions = (imageId: string) =>
-  queryOptions({
-    queryKey: ["image", imageId, "signed-url"] as const,
-    queryFn: () =>
-      httpClient.get<{ url: string; expiresAt: string }>(
-        `${STORY_API_BASE}/image/${imageId}/signed-url`,
-      ),
-    enabled: !!imageId,
-    staleTime: 55 * 60 * 1000,
-  });
