+++ frontend/src/features/scene-engine/steps/character/extraction/components/ReferenceImageViewer.tsx
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
 import { Skeleton } from "@scene-engine/components/Skeleton";
-import { imageSignedUrlOptions } from "../../character.queries";
+import { imageSignedUrlOptions } from "@scene-engine/shared/scene-engine.queries";
 import type { ImageRecord } from "@scene-engine/shared/scene-engine.types";
 
 type ReferenceImageViewerProps = {
