+++ frontend/src/features/scene-engine/steps/character/rendering/components/Carousel.tsx
@@ -1,7 +1,7 @@
 import { useRef, useState } from "react";
 import { useQuery, useQueryClient } from "@tanstack/react-query";
 import { Skeleton } from "@scene-engine/components/Skeleton";
-import { imageSignedUrlOptions } from "../../character.queries";
+import { imageSignedUrlOptions } from "@scene-engine/shared/scene-engine.queries";
 import type { ImageRecord } from "@scene-engine/shared/scene-engine.types";
 
 function CarouselImage({ render }: { render: ImageRecord }) {
